refactor: Separate logic from all files, create custom useHooks
