+++ src/screens/AddCategoryScreen/index.tsx
@@ -7,9 +7,8 @@ import {
   TouchableOpacity,
   View,
 } from 'react-native';
-import React, {useState} from 'react';
+import React from 'react';
 import {goBack} from '../../utils/navigationUtils';
-import useThemeColors from '../../hooks/useThemeColors';
 import AppHeader from '../../components/AppHeader';
 import CustomInput from '../../components/CustomInput';
 import PrimaryButton from '../../components/PrimaryButton';
@@ -17,50 +16,28 @@ import Icon from '../../components/Icons';
 import addTransactionStyles from '../AddTransactionsScreen/style';
 import allIcons from '../../../assets/jsons/categoryIcons.json';
 import allColors from '../../../assets/jsons/categoryColors.json';
-import {createCategory} from '../../services/CategoryService';
-import {useDispatch, useSelector} from 'react-redux';
-import {selectUserId} from '../../redux/slice/userIdSlice';
-import {FETCH_ALL_CATEGORY_DATA} from '../../redux/actionTypes';
 import styles from './style';
+import useAddCategory from './useAddCategory';
 
 const AddCategoryScreen = () => {
-  const colors = useThemeColors();
-  const dispatch = useDispatch();
-  const [categoryName, setCategoryName] = useState('');
-  const [isIconModalVisible, setIsIconModalVisible] = useState(false);
-  const [selectedIcon, setSelectedIcon] = useState('null');
-  const [isColorModalVisible, setIsColorModalVisible] = useState(false);
-  const [selectedColor, setSelectedColor] = useState('null');
-
-  const [searchText, setSearchText] = useState('');
-  const userId = useSelector(selectUserId);
-
-  const handleAddCategory = () => {
-    try {
-      createCategory(
-        categoryName,
-        Realm.BSON.ObjectID.createFromHexString(userId),
-        selectedIcon,
-        selectedColor,
-      );
-      dispatch({type: FETCH_ALL_CATEGORY_DATA});
-      goBack();
-    } catch (error) {
-      console.error('Error creating category:', error);
-    }
-  };
-
-  const handleIconModalClose = () => {
-    setIsIconModalVisible(false);
-  };
-
-  const handleColorModalClose = () => {
-    setIsColorModalVisible(false);
-  };
-
-  console.log(selectedColor);
-
-  console.log('this is the selectedicon', selectedIcon);
+  const {
+    colors,
+    categoryName,
+    setCategoryName,
+    selectedIcon,
+    setSelectedIcon,
+    isIconModalVisible,
+    setIsIconModalVisible,
+    isColorModalVisible,
+    setIsColorModalVisible,
+    handleAddCategory,
+    handleIconModalClose,
+    handleColorModalClose,
+    selectedColor,
+    setSelectedColor,
+    searchText,
+    setSearchText,
+  } = useAddCategory();
 
   const renderIcons = () => {
     const handleSelectIcon = iconName => {
