+++ src/screens/AddDebtorScreen/style.ts
@@ -1,0 +1,20 @@
+import {StyleSheet} from 'react-native';
+
+const addDebtorStyles = StyleSheet.create({
+  mainContainer: {
+    height: '100%',
+    paddingLeft: '6%',
+    paddingRight: '6%',
+  },
+  headerContainer: {
+    marginBottom: 20,
+    marginTop: 20,
+  },
+  labelText: {
+    fontFamily: 'FiraCode-Medium',
+    fontSize: 15,
+    includeFontPadding: false,
+  },
+});
+
+export default addDebtorStyles;
