+++ src/screens/SettingsScreen/index.tsx
@@ -1,166 +1,53 @@
 import {
-  Linking,
   Modal,
   ScrollView,
-  StyleSheet,
   Text,
   TextInput,
   TouchableOpacity,
   View,
   TouchableWithoutFeedback,
 } from 'react-native';
-import React, {useEffect, useState} from 'react';
-import useThemeColors from '../../hooks/useThemeColors';
-import {useDispatch, useSelector} from 'react-redux';
-import AsyncStorageService from '../../utils/asyncStorageService';
-import {
-  selectThemePreference,
-  setThemePreference,
-} from '../../redux/slice/themePreferenceSlice';
+import React from 'react';
 import CustomInput from '../../components/CustomInput';
 import PrimaryButton from '../../components/PrimaryButton';
-import {updateUserById} from '../../services/UserService';
-import {
-  selectCurrencyCode,
-  selectCurrencyId,
-  selectCurrencyName,
-  selectCurrencySymbol,
-  setCurrencyData,
-} from '../../redux/slice/currencyDataSlice';
-import {selectUserName, setUserName} from '../../redux/slice/userNameSlice';
-import {selectUserId} from '../../redux/slice/userIdSlice';
 import chooseCurrencyStyles from '../ChooseCurrencyScreen/style';
-import currencies from '../../../assets/jsons/currencies.json';
-import {updateCurrencyById} from '../../services/CurrencyService';
 import addCategoryStyles from '../AddCategoryScreen/style';
-import {getAppVersion} from '../../utils/getVersion';
 import Icon from '../../components/Icons';
 import {goBack} from '../../utils/navigationUtils';
+import useSettings from './useSettings';
+import styles from './style';
 
 const SettingsScreen = () => {
-  const userName = useSelector(selectUserName);
-  const userId = useSelector(selectUserId);
-  const currencyId = useSelector(selectCurrencyId);
-  const currencyCode = useSelector(selectCurrencyCode);
-  const currencyName = useSelector(selectCurrencyName);
-  const currencySymbol = useSelector(selectCurrencySymbol);
-  const selectedTheme = useSelector(selectThemePreference);
-
-  const currencyData = {
-    code: currencyCode,
-    name: currencyName,
-    symbol: currencySymbol,
-  };
-
-  const [isThemeModalVisible, setIsThemeModalVisible] = useState(false);
-  const [isNameModalVisible, setIsNameModalVisible] = useState(false);
-  const [name, setName] = useState(userName);
-  const [searchText, setSearchText] = useState('');
-  const [isCurrencyModalVisible, setIsCurrencyModalVisible] = useState(false);
-  const [selectedCurrency, setSelectedCurrency] = useState(currencyData);
-  const [filteredCurrencies, setFilteredCurrencies] = useState(currencies);
-  const appVersion = getAppVersion();
-
-  const colors = useThemeColors();
-  const dispatch = useDispatch();
-
-  useEffect(() => {
-    const getThemePreference = async () => {
-      try {
-        const theme = await AsyncStorageService.getItem('themePreference');
-        if (theme === null) {
-          dispatch(setThemePreference('system'));
-        } else {
-          dispatch(setThemePreference(theme));
-        }
-      } catch (error) {
-        console.error('Error getting theme preference:', error);
-      }
-    };
-
-    getThemePreference();
-  }, [selectedTheme]);
-
-  const handleThemeModalClose = () => {
-    setIsThemeModalVisible(false);
-  };
-
-  const handleNameModalClose = () => {
-    setIsNameModalVisible(false);
-  };
-
-  const handleCurrencyModalClose = () => {
-    setIsCurrencyModalVisible(false);
-  };
-
-  const handleCurrencySelect = currency => {
-    setSelectedCurrency(currency);
-  };
-
-  const handleThemeSelection = async theme => {
-    try {
-      if (theme === 'system') {
-        dispatch(setThemePreference(null));
-      } else {
-        dispatch(setThemePreference(theme));
-      }
-      setIsThemeModalVisible(false);
-      await AsyncStorageService.setItem('themePreference', theme);
-    } catch (error) {
-      console.error('Error saving theme preference:', error);
-    }
-  };
-
-  const handleNameUpdate = async () => {
-    try {
-      await updateUserById(Realm.BSON.ObjectID.createFromHexString(userId), {
-        username: name,
-      });
-      dispatch(setUserName(name));
-      setIsNameModalVisible(false);
-    } catch (error) {
-      console.error('Error updating the name:', error);
-    }
-  };
-
-  const handleSearch = (text: string) => {
-    setSearchText(text);
-    const filtered = currencies.filter(currency => {
-      const {code, name, symbol, symbolNative} = currency;
-      const searchItem = text.toLowerCase();
-
-      return (
-        code.toLowerCase().includes(searchItem) ||
-        name.toLowerCase().includes(searchItem) ||
-        symbol.toLowerCase().includes(searchItem) ||
-        symbolNative.toLowerCase().includes(searchItem)
-      );
-    });
-    setFilteredCurrencies(filtered);
-  };
-
-  const handleCurrencyUpdate = async () => {
-    try {
-      await updateCurrencyById(
-        Realm.BSON.ObjectID.createFromHexString(currencyId),
-        {
-          name: selectedCurrency.name,
-          code: selectedCurrency.code,
-          symbol: selectedCurrency.symbol,
-        },
-      );
-      const currencyData = {
-        currencyId,
-        currencyName: selectedCurrency.name,
-        currencySymbol: selectedCurrency.symbol,
-        currencyCode: selectedCurrency.code,
-      };
-      dispatch(setCurrencyData(currencyData));
-      setIsCurrencyModalVisible(false);
-    } catch (error) {
-      console.error('Error updating the currency:', error);
-    }
-  };
+  const {
+    isThemeModalVisible,
+    setIsThemeModalVisible,
+    isNameModalVisible,
+    setIsNameModalVisible,
+    name,
+    setName,
+    searchText,
+    isCurrencyModalVisible,
+    setIsCurrencyModalVisible,
+    filteredCurrencies,
+    appVersion,
+    colors,
+    handleThemeModalClose,
+    handleNameModalClose,
+    handleCurrencyModalClose,
+    handleCurrencySelect,
+    handleThemeSelection,
+    handleNameUpdate,
+    handleSearch,
+    handleCurrencyUpdate,
+    selectedCurrency,
+    selectedTheme,
+    userName,
+    currencySymbol,
+    currencyName,
+    handleRateNow,
+    handleGithub,
+    handlePrivacyPolicy,
+  } = useSettings();
 
   const renderRadioButtons = onThemeSelect => {
     const themes = ['light', 'dark', 'system'];
@@ -188,8 +75,6 @@ const SettingsScreen = () => {
       </TouchableOpacity>
     ));
   };
-
-  console.log(selectedCurrency);
 
   const renderCurrencySymbol = () => {
     return (
@@ -239,24 +124,6 @@ const SettingsScreen = () => {
     );
   };
 
-  const handleRateNow = () => {
-    console.log('rate on playstore');
-  };
-
-  const handleGithub = () => {
-    const githubRepoURL = 'https://github.com/indranilbhuin/zero';
-    Linking.openURL(githubRepoURL).catch(err =>
-      console.error('Error opening GitHub:', err),
-    );
-  };
-
-  const handlePrivacyPolicy = () => {
-    const privacyPolicyURL = 'https://github.com/indranilbhuin/zero/blob/master/PRIVACYPOLICY.md';
-    Linking.openURL(privacyPolicyURL).catch(err =>
-      console.error('Error opening GitHub:', err),
-    );
-  };
-
   return (
     <View
       style={[
@@ -266,7 +133,7 @@ const SettingsScreen = () => {
       <View style={styles.headerContainer}>
         <View style={styles.greetingsContainer}>
           <View style={styles.iconButtonContainer}>
-            <TouchableOpacity onPress={goBack}>
+            <TouchableOpacity onPress={() => goBack()}>
               <Icon
                 name="caret-back-circle"
                 size={25}
@@ -649,77 +516,3 @@ const SettingsScreen = () => {
 };
 
 export default SettingsScreen;
-
-const styles = StyleSheet.create({
-  mainContainer: {
-    height: '100%',
-    paddingLeft: '6%',
-    paddingRight: '6%',
-  },
-  headerContainer: {
-    flexDirection: 'row',
-    alignItems: 'center',
-    marginTop: '5%',
-    justifyContent: 'space-between',
-  },
-  titleText: {
-    fontFamily: 'FiraCode-Medium',
-    fontSize: 25,
-    includeFontPadding: false,
-  },
-  greetingsContainer: {
-    flexDirection: 'row',
-    alignItems: 'center',
-  },
-  settingsContainer: {
-    marginTop: 15,
-    borderWidth: 1,
-    borderRadius: 8,
-  },
-  individualSettingsContainer: {
-    flexDirection: 'row',
-    justifyContent: 'space-between',
-    height: 65,
-    alignItems: 'center',
-    padding: 10,
-    borderBottomWidth: 1,
-  },
-  modalContainer: {
-    width: '100%',
-    backgroundColor: 'rgba(0, 0, 0, 0.5)',
-    flex: 1,
-    justifyContent: 'flex-end',
-  },
-  subtitleText: {
-    fontFamily: 'FiraCode-Medium',
-    fontSize: 15,
-    includeFontPadding: false,
-  },
-  modal: {
-    borderTopLeftRadius: 15,
-    borderTopRightRadius: 15,
-    padding: 15,
-  },
-  radioButton: {
-    width: 20,
-    height: 20,
-    borderRadius: 10,
-    borderWidth: 2,
-    alignItems: 'center',
-    justifyContent: 'center',
-  },
-  radioButtonSelected: {
-    width: 10,
-    height: 10,
-    borderRadius: 5,
-  },
-  radioButtonContainer: {
-    flexDirection: 'row',
-    justifyContent: 'space-between',
-    alignItems: 'center',
-    marginBottom: 20,
-  },
-  iconButtonContainer: {
-    marginRight: 10,
-  },
-});
