+++ src/screens/SplashScreen/useSplash.ts
@@ -1,0 +1,16 @@
+import useThemeColors from '../../hooks/useThemeColors';
+import {navigate} from '../../utils/navigationUtils';
+
+const useSplash = () => {
+  const handleClick = () => {
+    navigate('PersonalizeScreen');
+  };
+
+  const colors = useThemeColors();
+  return {
+    handleClick,
+    colors,
+  };
+};
+
+export default useSplash;
