+++ src/screens/UpdateTransactionScreen/index.tsx
@@ -1,113 +1,48 @@
 import {
   ScrollView,
-  StyleSheet,
   Text,
   TextInput,
   TouchableOpacity,
   View,
 } from 'react-native';
-import React, {useState} from 'react';
+import React from 'react';
 import moment from 'moment';
 import AppHeader from '../../components/AppHeader';
 import CategoryContainer from '../../components/CategoryContainer';
 import CustomInput from '../../components/CustomInput';
 import PrimaryButton from '../../components/PrimaryButton';
 import SecondaryButton from '../../components/SecondaryButton';
-import {goBack, navigate} from '../../utils/navigationUtils';
+import {goBack} from '../../utils/navigationUtils';
 import Icon from '../../components/Icons';
 import addTransactionStyles from '../AddTransactionsScreen/style';
-import {RouteProp, useRoute} from '@react-navigation/native';
-import Category from '../../schemas/CategorySchema';
-import {useDispatch, useSelector} from 'react-redux';
-import {selectActiveCategories} from '../../redux/slice/categoryDataSlice';
-import {updateExpenseById} from '../../services/ExpenseService';
+import {useRoute} from '@react-navigation/native';
 import DateTimePicker from '@react-native-community/datetimepicker';
-import useThemeColors from '../../hooks/useThemeColors';
-import {selectCurrencySymbol} from '../../redux/slice/currencyDataSlice';
-import { getExpenseRequest } from '../../redux/slice/expenseDataSlice';
-
-type UpdateTransactionScreenRouteProp = RouteProp<
-  {
-    UpdateTransactionScreen: {
-      expenseId: string;
-      expenseTitle: string;
-      expenseDescription: string;
-      category: Category;
-      expenseDate: Date;
-      expenseAmount: string;
-    };
-  },
-  'UpdateTransactionScreen'
->;
+import useUpdateTransaction, {
+  UpdateTransactionScreenRouteProp,
+} from './useUpdateTransaction';
+import styles from './style';
 
 const UpdateTransactionScreen = () => {
   const route = useRoute<UpdateTransactionScreenRouteProp>();
-  const expenseData = route.params;
-  console.log('first', expenseData);
-  const colors = useThemeColors();
-  const [selectedCategories, setSelectedCategories] = useState([
-    expenseData?.category,
-  ]);
-  console.log('category', expenseData.expenseAmount);
-  const [expenseTitle, setExpenseTitle] = useState(expenseData.expenseTitle);
-  const [expenseDescription, setExpenseDescription] = useState(
-    expenseData.expenseDescription,
-  );
-  const [expenseAmount, setExpenseAmount] = useState(
-    String(expenseData.expenseAmount),
-  );
-  const [createdAt, setCreatedAt] = useState(expenseData.expenseDate);
-  const [showDatePicker, setShowDatePicker] = useState(false);
-  const categories = useSelector(selectActiveCategories);
-  const currencySymbol = useSelector(selectCurrencySymbol);
-  const dispatch = useDispatch();
-
-  const handleDateChange = (event, selectedDate) => {
-    const currentDate = selectedDate || createdAt;
-    const utcDate = moment(currentDate).utc().toDate();
-    setCreatedAt(utcDate);
-    setShowDatePicker(false);
-  };
-
-  const toggleCategorySelection = category => {
-    if (selectedCategories.includes(category)) {
-      setSelectedCategories([]);
-    } else {
-      setSelectedCategories([category]);
-    }
-  };
-
-  const handleAddCategory = () => {
-    navigate('CategoryScreen');
-  };
-
-  const handleUpdateExpense = () => {
-    if (
-      expenseTitle.trim() === '' ||
-      expenseAmount === null ||
-      selectedCategories.length === 0
-    ) {
-      return;
-    }
-    const categoryId = selectedCategories[0]._id;
-    try {
-      updateExpenseById(
-        Realm.BSON.ObjectID.createFromHexString(expenseData.expenseId),
-        categoryId,
-        expenseTitle,
-        Number(expenseAmount),
-        expenseDescription,
-        createdAt,
-      );
-      
-      dispatch(getExpenseRequest());
-      goBack();
-    } catch (error) {
-      console.error('Error updating expense:', error);
-    }
-  };
-
-  console.log(selectedCategories);
+  const {
+    colors,
+    selectedCategories,
+    expenseTitle,
+    setExpenseTitle,
+    expenseDescription,
+    setExpenseDescription,
+    expenseAmount,
+    setExpenseAmount,
+    createdAt,
+    showDatePicker,
+    setShowDatePicker,
+    categories,
+    currencySymbol,
+    handleDateChange,
+    toggleCategorySelection,
+    handleAddCategory,
+    handleUpdateExpense,
+  } = useUpdateTransaction(route);
 
   return (
     <View
@@ -249,29 +184,3 @@ const UpdateTransactionScreen = () => {
 };
 
 export default UpdateTransactionScreen;
-
-const styles = StyleSheet.create({
-  textInputContainer: {
-    height: 60,
-    alignItems: 'center',
-    borderWidth: 2,
-    marginTop: 5,
-    marginBottom: 15,
-    borderRadius: 15,
-    paddingLeft: 20,
-    justifyContent: 'flex-start',
-    flexDirection: 'row',
-  },
-  textInput: {
-    padding: 20,
-    height: 60,
-    width: '100%',
-    fontFamily: 'FiraCode-Medium',
-    includeFontPadding: false,
-  },
-  labelText: {
-    fontFamily: 'FiraCode-Medium',
-    fontSize: 15,
-    includeFontPadding: false,
-  },
-});
